Java program that reads, processes, and displays attendance records in a Zoom meeting report provided as a txt file.
The first line of the input file is a header line which includes title of each column of the data file:
 Join Time: Date and time when participants join the meeting
 Leave Time: Date and time when participants leave the meeting
 Duration: Duration (in terms of minutes) passed between join and leave time
 Name: Name & surname
 Email: Email address
Each of the remaining lines includes a record for a connection of an attendee to the meeting where attendees are uniquely recognized by their name & surname. Name of an attendee might include more than one word but surname is always a single word at the end of name & surname.

Name & surname refer to the same attendee if they are the same without considering uppercase/lowercase distinction of the letters. For example:
Eda Nur Yilmaz and EDA NUR YILMAZ refer to the same attendee.
Eda Nur Yilmaz and Eda Nur YILMAZ refer to the same attendee.
Of course, order and number of the words in name & surname matters. For example:
Eda Nur Yilmaz and Nur Eda Yilmaz refer to different attendees.
EDA NUR YILMAZ and EDA YILMAZ refer to different attendees.
Assume that there is a single space character between words in name & surname column of input file. For example:
EDA NUR YILMAZ and EDANUR YILMAZ refer to different attendees.
Of course, if there is a difference between surnames or names, they refer to different attendees. For example:
EDANUR YILMAZ and EDA YILMAZ refer to different attendees.
EDA YILMAZ and EDA YILMAZCAN refer to different attendees.
Notice that it is possible that e-mail address is not provided in some records.
Also, notice that there might be more than one record for some of attendees as they leave and join the meeting multiple times. Assume that there is no time overlap between multiple connections made by each attendee. That means each attendee has at most one connection at each moment of the meeting.
Note: The number of test cases containing multiple connections situation will be less than the others, if you do not do this part, you will not get full points.
Assume that duration is always provided for each record and it is compatible with the difference between join and leave time so you will consider duration of each connection as the number at the middle of its corresponding line (you do not need to compute duration of each connection from join and leave time).
Assume that the number of attendees in a Zoom session is not greater than 500.